+++ lib/expressions/value_expression.rb
@@ -1,16 +1,18 @@
 # frozen_string_literal: true
 
-# A value expression presents a type as a value:
-# 1. It is an instance
-# 2. It has a class method
-class ValueExpression
-  attr_reader :value
+module Low
+  # A value expression presents a type as a value:
+  #   1. It is an instance
+  #   2. It mimics the class method
+  class ValueExpression
+    attr_reader :value
 
-  def initialize(value:)
-    @value = value
-  end
+    def initialize(value:)
+      @value = value
+    end
 
-  def class
-    @value
+    def class
+      @value
+    end
   end
 end